ml-backend/requirements.txt
anyio==4.12.0
av==16.0.1
blinker==1.9.0
breadability==0.1.20
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.1
decorator==5.2.1
dnspython==2.8.0
docopt==0.6.2
exceptiongroup==1.3.1
faster-whisper==1.2.1
ffmpeg-python==0.2.0
filelock==3.20.0
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==25.9.23
fsspec==2025.10.0
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.6
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.2
MarkupSafe==3.0.3
moviepy==2.2.1
mpmath==1.3.0
nltk==3.9.2
numpy==1.26.4
onnxruntime==1.23.2
opencv-python==4.10.0.82
packaging==25.0
pillow==11.3.0
proglog==0.1.12
protobuf==6.33.1
pycountry==24.6.1
pydub==0.25.1
pymongo==4.15.4
pyreadline3==3.5.4
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
shellingham==1.5.4
SpeechRecognition==3.14.4
sumy==0.11.0
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
